+++ src/app/(providers)/(root)/_components/DashBoardHeader/index.ts
@@ -1,0 +1,1 @@
+export { default } from './DashBoardHeader';
